+++ kernel/drivers/timer.c
@@ -2,7 +2,7 @@
 
 void timer_sleep(unsigned int ms)
 {
-    unsigned int start = timer_ticks();
+    unsigned long start = timer_ticks();
 
     while (timer_ticks() < start + (ms * 1000))
         ;
